nvim: slightly improved man page highlighting?
It still needs a lot of work. The way syntax/man.vim and man.lua work
together is really cursed, and they are very inflexible. man.lua is
applying manBold, manItalic, etc. in places where the man syntax file is
trying to display C syntax highlighting, which is really irritating.